Extract DataStore from AndroidLocalStorage for easier testing
